Re-export `FileUtil` inside `index.js`, fix import for `ErrorAlert`
