+++ lib/util/Alert/index.js
@@ -1,4 +1,4 @@
-const { ErrorAlert } = importModule('ErrorAlert');
+const { ErrorAlert } = importModule('./ErrorAlert');
 
 module.exports = {
 	ErrorAlert,
