+++ lib/util/FileUtil/FileUtil.test.js
@@ -1,6 +1,12 @@
+// If running outside the Scriptable environment...
+if (typeof config === 'undefined') {
+	// ... overwrite the `importModule()` global function.
+	global.importModule = modulePath => require(modulePath)
+}
+
 const { ImageType } = require('../ImageUtil')
 
-const { FileUtil } = require('.')
+const { FileUtil } = require('./FileUtil')
 
 describe('FileUtil', () => {
 	describe('when finding image by name', () => {
